clean up tests for registry clients, rename registry utils
