+++ test/registry/FHIRRegistryClient.test.ts
@@ -63,7 +63,7 @@ describe('FHIRRegistryClient', () => {
   });
 
   describe('#download', () => {
-    describe('#downloadInvalidVersion', () => {
+    describe('#downloadSpecificVersion', () => {
       beforeEach(() => {
         loggerSpy.reset();
       });
@@ -71,15 +71,33 @@ describe('FHIRRegistryClient', () => {
         axiosSpy = jest.spyOn(axios, 'get').mockImplementation((uri: string): any => {
           if (uri === 'https://packages.fhir.org/hl7.terminology.r4') {
             return { data: TERM_PKG_RESPONSE };
+          } else if (uri === 'https://packages.fhir.org/hl7.terminology.r4/1.2.3-test') {
+            return {
+              status: 200,
+              data: Readable.from(['1.2.3-test-data'])
+            };
+          } else if (uri === 'https://packages.fhir.org/hl7.terminology.r4/2.2.2') {
+            return {
+              status: 200,
+              data: ''
+            };
+          } else if (uri === 'https://packages.fhir.org/hl7.terminology.r4/3.3.3') {
+            return {
+              status: 200
+            };
+          } else if (uri === 'https://packages.fhir.org/hl7.terminology.r4/5.5.5') {
+            return {
+              status: 'wrong-type',
+              data: Readable.from(['1.2.4-no-manifest-test-data'])
+            };
+          } else if (uri === 'https://packages.fhir.org/hl7.terminology.r4/5.5.6-test') {
+            return {
+              data: Readable.from(['5.5.6-test-data'])
+            };
           } else if (uri === 'https://packages.fhir.org/hl7.terminology.r4/1.1.2') {
             return {
               status: 404,
               data: Readable.from(['1.1.2-test-data'])
-            };
-          } else if (uri === 'https://packages.fhir.org/hl7.terminology.r4/1.1.1') {
-            return {
-              status: 200,
-              data: null
             };
           } else {
             throw new Error('Not found');
@@ -89,6 +107,44 @@ describe('FHIRRegistryClient', () => {
 
       afterAll(() => {
         axiosSpy.mockRestore();
+      });
+
+      it('should throw error if no name given for download method', async () => {
+        const latest = client.download('', '5.5.5');
+        await expect(loggerSpy.getLastMessage('info')).toBe(
+          'Attempting to download #5.5.5 from https://packages.fhir.org//5.5.5'
+        );
+        await expect(latest).rejects.toThrow(Error);
+        await expect(latest).rejects.toThrow('Not found');
+      });
+
+      it('should throw error if no version given for download method', async () => {
+        const latest = client.download('hl7.terminology.r4', '');
+        await expect(latest).rejects.toThrow(Error);
+        await expect(latest).rejects.toThrow('Not found');
+        await expect(loggerSpy.getLastMessage('info')).toBe(
+          'Attempting to download hl7.terminology.r4# from https://packages.fhir.org/hl7.terminology.r4/'
+        );
+      });
+
+      it('should throw error if no endpoint given for download method', async () => {
+        const emptyClient = new FHIRRegistryClient('', { log: loggerSpy.log });
+        const latest = emptyClient.download('hl7.terminology.r4', '1.2.3-test');
+        await expect(latest).rejects.toThrow(Error);
+        await expect(latest).rejects.toThrow('Not found');
+        await expect(loggerSpy.getLastMessage('info')).toBe(
+          'Attempting to download hl7.terminology.r4#1.2.3-test from /hl7.terminology.r4/1.2.3-test'
+        );
+      });
+
+      it('should get the data of the package when 200 response', async () => {
+        const latest = await client.download('hl7.terminology.r4', '1.2.3-test');
+        expect(loggerSpy.getLastMessage('info')).toBe(
+          'Attempting to download hl7.terminology.r4#1.2.3-test from https://packages.fhir.org/hl7.terminology.r4/1.2.3-test'
+        );
+        expect(loggerSpy.getAllMessages('error')).toHaveLength(0);
+        expect(latest).toBeInstanceOf(Readable);
+        expect(latest.read()).toBe('1.2.3-test-data');
       });
 
       it('should throw error when trying to get the version of a package on the packages server but status is not 200', async () => {
@@ -102,139 +158,32 @@ describe('FHIRRegistryClient', () => {
         );
       });
 
-      it('should throw error when trying to get the version of a package on the packages server but returns no data', async () => {
-        // Note: don't know of a scenario where this would occur but testing for completeness.
-        const latest = client.download('hl7.terminology.r4', '1.1.1');
-        expect(loggerSpy.getLastMessage('info')).toBe(
-          'Attempting to download hl7.terminology.r4#1.1.1 from https://packages.fhir.org/hl7.terminology.r4/1.1.1'
-        );
-        await expect(latest).rejects.toThrow(Error);
+      it('should throw error when trying to get the version of a package on the server but returns status with incorrect type', async () => {
+        const latest = client.download('hl7.terminology.r4', '5.5.5');
         await expect(latest).rejects.toThrow(
-          'Failed to download hl7.terminology.r4#1.1.1 from https://packages.fhir.org/hl7.terminology.r4/1.1.1'
+          'Failed to download hl7.terminology.r4#5.5.5 from https://packages.fhir.org/hl7.terminology.r4/5.5.5'
         );
       });
-    });
 
-    describe('#downloadSpecificVersion', () => {
-      describe('#CreateURL', () => {
-        beforeEach(() => {
-          loggerSpy.reset();
-        });
-        beforeAll(() => {
-          axiosSpy = jest.spyOn(axios, 'get').mockImplementation((uri: string): any => {
-            if (uri === 'https://packages.fhir.org/hl7.terminology.r4') {
-              return { data: TERM_PKG_RESPONSE };
-            } else {
-              throw new Error('Not found');
-            }
-          });
-        });
-
-        afterAll(() => {
-          axiosSpy.mockRestore();
-        });
-
-        it('should throw error if no name given for download method', async () => {
-          const latest = client.download('', '5.5.5');
-          // no message logged
-          await expect(latest).rejects.toThrow(Error);
-          await expect(latest).rejects.toThrow('Not found');
-        });
-
-        it('should throw error if no version given for download method', async () => {
-          const latest = client.download('hl7.terminology.r4', '');
-          // no message logged
-          await expect(latest).rejects.toThrow(Error);
-          await expect(latest).rejects.toThrow('Not found');
-        });
-
-        it('should throw error if no endpoint given for download method', async () => {
-          const emptyClient = new FHIRRegistryClient('', { log: loggerSpy.log });
-          const latest = emptyClient.download('hl7.terminology.r4', '1.2.3-test');
-          // no message logged
-          await expect(latest).rejects.toThrow(Error);
-          await expect(latest).rejects.toThrow('Not found');
-        });
+      it('should throw error when trying to get the version of a package on the server but returns no status', async () => {
+        const latest = client.download('hl7.terminology.r4', '5.5.6-test');
+        await expect(latest).rejects.toThrow(
+          'Failed to download hl7.terminology.r4#5.5.6-test from https://packages.fhir.org/hl7.terminology.r4/5.5.6-test'
+        );
       });
 
-      describe('#DownloadFromURL', () => {
-        beforeEach(() => {
-          loggerSpy.reset();
-        });
-        beforeAll(() => {
-          axiosSpy = jest.spyOn(axios, 'get').mockImplementation((uri: string): any => {
-            if (uri === 'https://packages.fhir.org/hl7.terminology.r4') {
-              return { data: TERM_PKG_RESPONSE };
-            } else if (uri === 'https://packages.fhir.org/hl7.terminology.r4/1.2.3-test') {
-              return {
-                status: 200,
-                data: Readable.from(['1.2.3-test-data'])
-              };
-            } else if (uri === 'https://packages.fhir.org/hl7.terminology.r4/2.2.2') {
-              return {
-                status: 200,
-                data: ''
-              };
-            } else if (uri === 'https://packages.fhir.org/hl7.terminology.r4/3.3.3') {
-              return {
-                status: 200
-              };
-            } else if (uri === 'https://packages.fhir.org/hl7.terminology.r4/5.5.5') {
-              return {
-                status: 'wrong-type',
-                data: Readable.from(['1.2.4-no-manifest-test-data'])
-              };
-            } else if (uri === 'https://packages.fhir.org/hl7.terminology.r4/5.5.6-test') {
-              return {
-                data: Readable.from(['5.5.6-test-data'])
-              };
-            } else {
-              throw new Error('Not found');
-            }
-          });
-        });
+      it('should throw error when trying to get the version of a package on the server but returns 200 status and data of incorrect type', async () => {
+        const latest = client.download('hl7.terminology.r4', '2.2.2');
+        await expect(latest).rejects.toThrow(
+          'Failed to download hl7.terminology.r4#2.2.2 from https://packages.fhir.org/hl7.terminology.r4/2.2.2'
+        );
+      });
 
-        afterAll(() => {
-          axiosSpy.mockRestore();
-        });
-
-        it('should get the data of the package when 200 response', async () => {
-          const latest = await client.download('hl7.terminology.r4', '1.2.3-test');
-          expect(loggerSpy.getLastMessage('info')).toBe(
-            'Attempting to download hl7.terminology.r4#1.2.3-test from https://packages.fhir.org/hl7.terminology.r4/1.2.3-test'
-          );
-          expect(loggerSpy.getAllMessages('error')).toHaveLength(0);
-          expect(latest).toBeInstanceOf(Readable);
-          expect(latest.read()).toBe('1.2.3-test-data');
-        });
-
-        it('should throw error when trying to get the version of a package on the server but returns 200 status and data of incorrect type', async () => {
-          const latest = client.download('hl7.terminology.r4', '2.2.2');
-          await expect(latest).rejects.toThrow(
-            'Failed to download hl7.terminology.r4#2.2.2 from https://packages.fhir.org/hl7.terminology.r4/2.2.2'
-          );
-        });
-
-        it('should throw error when trying to get the version of a package on the server but returns 200 status and no data field', async () => {
-          const latest = client.download('hl7.terminology.r4', '3.3.3');
-          await expect(latest).rejects.toThrow(
-            'Failed to download hl7.terminology.r4#3.3.3 from https://packages.fhir.org/hl7.terminology.r4/3.3.3'
-          );
-        });
-
-        it('should throw error when trying to get the version of a package on the server but returns status with incorrect type', async () => {
-          const latest = client.download('hl7.terminology.r4', '5.5.5');
-          await expect(latest).rejects.toThrow(
-            'Failed to download hl7.terminology.r4#5.5.5 from https://packages.fhir.org/hl7.terminology.r4/5.5.5'
-          );
-        });
-
-        it('should throw error when trying to get the version of a package on the server but returns no status', async () => {
-          const latest = client.download('hl7.terminology.r4', '5.5.6-test');
-          await expect(latest).rejects.toThrow(
-            'Failed to download hl7.terminology.r4#5.5.6-test from https://packages.fhir.org/hl7.terminology.r4/5.5.6-test'
-          );
-        });
+      it('should throw error when trying to get the version of a package on the server but returns 200 status and no data field', async () => {
+        const latest = client.download('hl7.terminology.r4', '3.3.3');
+        await expect(latest).rejects.toThrow(
+          'Failed to download hl7.terminology.r4#3.3.3 from https://packages.fhir.org/hl7.terminology.r4/3.3.3'
+        );
       });
     });
